% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{TwoStepCLogit-package}
\alias{TwoStepCLogit}
\alias{TwoStepCLogit-package}
\title{Conditional logistic regression: A two-step estimation method}
\description{
Conditional logistic regression with longitudinal follow up and
individual-level random coefficients: A stable and efficient two-step
estimation method (see \code{\link{Ts.estim}}).
}
\details{
\tabular{ll}{
Package: \tab TwoStepCLogit\cr
Type: \tab Package\cr
Version: \tab 1.2.3\cr
Date: \tab 2014-10-23\cr
License: \tab GPL-2\cr
}
}
\author{
Radu V. Craiu, Thierry Duchesne, Daniel Fortin and Sophie
Baillargeon

Maintainer: Thierry Duchesne <thierry.duchesne@mat.ulaval.ca>
}
\references{
Craiu, R.V., Duchesne, T., Fortin, D. and Baillargeon, S.
(2011), Conditional Logistic Regression with Longitudinal Follow-up and
Individual-Level Random Coefficients: A Stable and Efficient Two-Step
Estimation Method, \emph{Journal of Computational and Graphical Statistics}. \bold{20}(3), 767-784.
}
\keyword{package}

